zeta9: dtype dispatch on f for sidecar root storage
When f ≤ 15, store roots_flat as int32 instead of int64 in stage 3
(find_roots_exact_v2) and stage 4 (build_phase_sidecar_binned_mpi).
bin_off changed to int32 unconditionally (bounded by per-Y root count).
At f ≥ 16 dispatch auto-flips to int64 to preserve correctness at the
ε=10⁻¹⁰ target (f≈21).

Savings: ~5-20 GB resident RAM at f=6 mpi=32 node (the sidecar mmap
is the biggest artifact). Speed neutral or slightly faster (half the
bytes through page cache).

Single source of truth: pick_root_dtype(f) in find_roots_exact_v2.py.
Threshold = 15, set by _ROOT_DTYPE_F_MAX_INT32.

Write-time assertion in _pack_roots_flat catches any value that would
silently truncate. Backward-compatible with existing int64 caches:
stage 4/5 readers inherit whatever dtype was saved.

End-to-end verified at θ=0.5 ε=0.1 max-f=1 (frob=0.0398, reproduced)
and θ=π/4 ε=0.1 max-f=1 (frob=0.0475). Sidecar roots_flat = int32,
root values in [-12, 12] at f=2 — well under int32 limit.

+++ zeta9/zeta9/build_phase_sidecar_binned_mpi.py
@@ -80,14 +80,19 @@ def _existing_sidecar_ok(meta_path: str, n_phase_bins: int) -> dict | None:
 def _build_one_batch(root_path: str, sidecar_dir: str, batch_idx: int, n_phase_bins: int) -> dict:
     data = np.load(root_path, allow_pickle=True, mmap_mode='r')
     Y = np.asarray(data["Y"], dtype=np.int64)
-    roots_flat = np.asarray(data["roots_flat"], dtype=np.int64)
+    # Preserve stage-3's dtype on roots_flat (int32 when f≤15, int64 otherwise).
+    # The .npy header records the dtype; we inherit it for the sidecar so
+    # mmap-resident pages on stage 5 are half-size when int32.
+    roots_flat = np.asarray(data["roots_flat"])
     roots_off = np.asarray(data["roots_off"], dtype=np.int64)
 
-    out_roots = np.empty_like(roots_flat)
+    out_roots = np.empty_like(roots_flat)  # inherits roots_flat dtype
     out_phases = np.empty((roots_flat.shape[0],), dtype=np.float64)
     out_z_re = np.empty((roots_flat.shape[0],), dtype=np.float64)
     out_z_im = np.empty((roots_flat.shape[0],), dtype=np.float64)
-    bin_off = np.zeros((Y.shape[0], int(n_phase_bins) + 1), dtype=np.int64)
+    # bin_off values are cumulative per-bin root counts, bounded by per-Y root
+    # count (~thousands at any f). int32 is plenty and halves resident pages.
+    bin_off = np.zeros((Y.shape[0], int(n_phase_bins) + 1), dtype=np.int32)
 
     for row in range(Y.shape[0]):
         a = int(roots_off[row]); b = int(roots_off[row + 1])
@@ -109,7 +114,8 @@ def _build_one_batch(root_path: str, sidecar_dir: str, batch_idx: int, n_phase_b
         local_bins = np.floor(phases / TWOPI * n_phase_bins).astype(np.int64)
         local_bins = np.clip(local_bins, 0, n_phase_bins - 1)
         counts = np.bincount(local_bins, minlength=n_phase_bins)
-        bin_off[row, 1:] = np.cumsum(counts, dtype=np.int64)
+        # int32 sum is safe: bounded by per-Y root count (~thousands)
+        bin_off[row, 1:] = np.cumsum(counts, dtype=np.int32)
 
     bp = _sidecar_batch_paths(sidecar_dir, batch_idx)
     np.save(bp["roots_flat"], out_roots)
